requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
PyMuPDF
sentence-transformers
faiss-cpu
groq
